tools/social-poster/requirements.txt
flask==3.1.*
gunicorn==23.*
atproto>=0.0.55
tweepy>=4.14
requests>=2.31
python-dotenv>=1.0
beautifulsoup4>=4.12
